+++ compiler/src/main/kotlin/com/squareup/metro/extensions/SquareMetroExtensionsPluginComponentRegistrar.kt
@@ -2,6 +2,7 @@ package com.squareup.metro.extensions
 
 import com.fueledbycaffeine.autoservice.AutoService
 import com.squareup.metro.extensions.featureflag.ContributesFeatureFlagIrExtension
+import com.squareup.metro.extensions.robot.ContributesRobotIrExtension
 import com.squareup.metro.extensions.service.ContributesServiceIrExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
@@ -20,6 +21,7 @@ public class SquareMetroExtensionsPluginComponentRegistrar : CompilerPluginRegis
     FirExtensionRegistrarAdapter.registerExtension(
       SquareMetroExtensionsPluginRegistrar(isReleaseBuild, enableDevelopmentAppComponent)
     )
+    IrGenerationExtension.registerExtension(ContributesRobotIrExtension())
     IrGenerationExtension.registerExtension(ContributesServiceIrExtension())
     IrGenerationExtension.registerExtension(ContributesFeatureFlagIrExtension())
   }