+++ compiler/src/test/resources/dump/contributesrobot/contributesRobot.kt
@@ -4,4 +4,7 @@ package com.test
 import com.squareup.anvil.extension.ContributesRobot
 import com.squareup.instrumentation.robots.ScreenRobot
 
-@Inject @ContributesRobot(Unit::class) class AbcRobot : ScreenRobot<AbcRobot>()
+class Dependency @Inject constructor()
+
+@Inject @ContributesRobot(Unit::class)
+class AbcRobot(val dependency: Dependency) : ScreenRobot<AbcRobot>()